+++ oneflow/user/kernels/triu_kernel.cu
@@ -68,8 +68,8 @@ __global__ void TriuWarpProcessRowGpu<half>(const int64_t total_rows, const int6
       const half2 x_val = x_h2[idx];
       half2 y_val;
 #ifdef WITH_ROCM
-      y_val.data.x = (2 * col) < row + diagonal ? static_cast<half>(0) : x_val.data.x;
-      y_val.data.y = (2 * col + 1) < row + diagonal ? static_cast<half>(0) : x_val.data.y;
+      y_val.data.x = (2 * col) < row + diagonal ? static_cast<half>(0) : static_cast<half>(x_val.data.x);
+      y_val.data.y = (2 * col + 1) < row + diagonal ? static_cast<half>(0) : static_cast<half>(x_val.data.y);
 #else
       y_val.x = (2 * col) < row + diagonal ? static_cast<half>(0) : x_val.x;
       y_val.y = (2 * col + 1) < row + diagonal ? static_cast<half>(0) : x_val.y;
